app/primer_cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "primer-cli"
version = "0.1.0"
description = "CLI pipeline: NCBI -> MAFFT -> conserved regions"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Your Name", email = "you@example.com" }
]
keywords = ["bioinformatics", "ncbi", "mafft", "alignment", "conserved-regions"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
  "requests>=2.31",
  "biopython>=1.83",
  "numpy>=1.26",
  "scikit-bio>=0.6.2",
  "primer3-py>=2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.3",
  "mypy>=1.8",
  "black>=24.0",
  "pre-commit>=3.6",
]
notebook = [
  "jupyterlab>=4.0",
  "ipykernel>=6.0",
  "jupytext>=1.16",
  "pandas>=2.0",
  "scipy>=1.11",
  "matplotlib>=3.8",
]

[project.scripts]
primer-cli = "primer_cli.cli.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["primer_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "integration: slower end-to-end checks over fixture datasets",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
